Change type of shiftDatabase param to size_t

+++ main.c
@@ -224,7 +224,7 @@ void displayStudent() {
 void searchStudent(int id) {
 }
 
-void shiftDatabase(int row) {
+void shiftDatabase(size_t row) {
     for (int i = row+1; i < entries; i++) {
         for (int j = 0; j < 6; j++) {
             STUDENT_DATABASE[i-1][j] = STUDENT_DATABASE[i][j];
